+++ contracts/omniflix-profile/src/contract.rs
@@ -1,13 +1,15 @@
 use cosmwasm_std::{
-    entry_point, to_binary, Deps, DepsMut, Env, MessageInfo, Response,
+    entry_point, to_binary, Deps, DepsMut, Env, MessageInfo, Response, Coin, Decimal, Uint128,
     StdResult, WasmMsg,
 };
-use omniflix_std::types::omniflix::onft::v1::{MsgCreateDenom, MsgMintONFT, Metadata};
+use omniflix_std::types::omniflix::onft::v1beta1::{MsgMintOnft, MsgCreateDenom, Metadata, WeightedAddress};
 
 use crate::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
-use crate::state::{Config, Profile, CONFIG, PROFILES};
+use crate::state::{CONFIG, PROFILES, PROFILE_SEQ, Config, Profile};
 use crate::error::ContractError;
-use rand::{distributions::Alphanumeric, Rng};
+use serde_json::json;
+
+use std::str::FromStr;
 
 #[entry_point]
 pub fn instantiate(
@@ -25,6 +27,13 @@ pub fn instantiate(
 
     CONFIG.save(deps.storage, &config)?;
 
+    let creation_fee = Coin {
+        denom: "uflix".to_string(),
+        amount: Uint128::from_str("1000000")?,
+    };
+
+    let royalty_receivers:Vec<WeightedAddress> = Vec::new();
+
     let create_denom_msg = MsgCreateDenom {
         id: msg.denom_id,
         name: msg.denom_name,
@@ -41,6 +50,8 @@ pub fn instantiate(
         uri: "".into(),
         uri_hash: "".into(),
         data: "".into(),
+        creation_fee: Some(creation_fee.into()),
+        royalty_receivers,
     };
 
     Ok(Response::new()
@@ -90,26 +101,30 @@ pub fn execute_create_profile(
     PROFILE_SEQ.save(deps.storage, &seq)?;
 
     // Generate NFT ID with sequence
-    let nft_id = format!("ofp", seq);
+    let nft_id = format!("ofp{}", seq);
 
     let metadata = Metadata {
         name: username.clone(),
         description: bio.clone(),
-        uri: profile_image.clone(),
-        uri_hash: "".into(),
+        media_uri: "_".to_string(),
+        preview_uri: "_".to_string(),
+        uri_hash: "_".to_string(),
+    };
+
+    let mint_onft_msg = MsgMintOnft {
+        id: nft_id.clone(),
+        denom_id: config.denom_id,
+        metadata: Some(metadata),
         data: serde_json::json!({
             "username": username,
             "bio": bio,
             "social_links": social_links,
             "profile_image": profile_image
         }).to_string(),
-    };
-
-    let mint_onft_msg = MsgMintONFT {
-        id: nft_id.clone(),
-        denom_id: config.denom_id,
-        metadata: Some(metadata),
         transferable: false,
+        extensible: true,
+        nsfw: false,
+        royalty_share: Decimal::from_str("0.01")?.to_string(),
         sender: env.contract.address.to_string(),
         recipient: info.sender.to_string(),
     };
@@ -119,7 +134,7 @@ pub fn execute_create_profile(
         bio,
         social_links,
         profile_image,
-        nft_id,
+        nft_id: nft_id.clone(),
     })?;
 
     Ok(Response::new()
