+++ contracts/omniflix-profile/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod contract;
+pub mod error;
+pub mod msg;
+pub mod state;
