+++ contracts/omniflix-profile/tests/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod integration_test;
